switch to using local outputs
previously the wrapper script used to use files created in memory
(/tmp) to use as the basis for invoking `docker compose` which led
to issues with regards to other related files such as env files
which needed to be linked, it also did not make it straitforward
to carry maintenance commands such as logs and also to attend to
usecases where the ports did not need to change (such as in the
case of using tunnels to local services which are pinned to speci-
fic port numbers). For all these reasons and others I decided to
switch to creating files with a leading tilde ~ in the name which
act as living memory in the folder and pinning the port numbers
while also allowing easily for calling maintenance commands.